% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{auto_choose_clusters}
\alias{auto_choose_clusters}
\title{Cluster Pathways and Automatically Cut the Dendrogram}
\usage{
auto_choose_clusters(result_df, agg_method = "average",
  plot_heatmap = FALSE, plot_dend = FALSE)
}
\arguments{
\item{result_df}{resulting data frame of enriched pathways from the wrapper function
\code{run_pathfindR}. Columns are: \enumerate{
 \item{ID}{KEGG ID of the enriched pathway}
 \item{Pathway}{Description of the enriched pathway}
 \item{Fold_Enrichment}{Fold enrichment value for the enriched pathway}
 \item{occurrence}{the number of iterations that the given pathway was found to enriched over all iterations}
 \item{lowest_p}{the lowest adjusted-p value of the given pathway over all iterations}
 \item{highest_p}{the highest adjusted-p value of the given pathway over all iterations}
 \item{Up_regulated}{the up-regulated genes in the input involved in the given pathway, comma-separated}
 \item{Down_regulated}{the down-regulated genes in the input involved in the given pathway, comma-separated}}}

\item{agg_method}{the agglomeration method to be used if plotting heatmap
(see next argument, default: average).}

\item{plot_heatmap}{boolean value indicating whether or not to plot the heat
map of pathway clustering (default: FALSE).}

\item{plot_dend}{boolean value indicating whether or not to plot the dendrogram
partitioned into the optimal number of clusters, shown by rectangles (default: FALSE)}
}
\value{
The function returns a data frame consisting of 10 columns: \describe{
  \item{ID}{KEGG ID of the enriched pathway}
  \item{Pathway}{Description of the enriched pathway}
  \item{Fold_Enrichment}{Fold enrichment value for the enriched pathway}
  \item{occurrence}{the number of iterations that the given pathway was found to enriched over all iterations}
  \item{lowest_p}{the lowest adjusted-p value of the given pathway over all iterations}
  \item{highest_p}{the highest adjusted-p value of the given pathway over all iterations}
  \item{Up_regulated}{the up-regulated genes in the input involved in the given pathway, comma-separated}
  \item{Down_regulated}{the down-regulated genes in the input involved in the given pathway, comma-separated}
  \item{Cluster}{the cluster to which the pathway is assigned}
  \item{Status}{whether the pathway is the "Representative" pathway in its cluster or only a "Member"}
}
}
\description{
This function first calculates the pairwise distances between the
pathways in the \code{result_df} data frame. Next, the average silhouette
width for each possible number of clusters are calculated. The optimal number
of clusters is selected as the one with the highest average silhouette width.
The dendrogram is cut into this number of clusters, and the pathways with the
lowest p value within each cluster are chosen as representative pathways. See
"Chen, Y. A. et al. Integrated pathway clusters with coherent biological
themes for target prioritisation. PLoS One 9, e99030,
doi:10.1371/journal.pone.0099030 (2014)." for details on the method of
pathway clustering.
}
\examples{
auto_choose_clusters(RA_output)
}
\seealso{
Use \code{\link{choose_clusters}} for manual selection of clusters.
  See \code{\link{calculate_pwd}} for calculation of pairwise
  distances between enriched pathways. See \code{\link{run_pathfindR}}
  for the wrapper function of the pathfindR enrichment workflow.
}
