% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biocarta_descriptions}
\alias{biocarta_descriptions}
\title{BioCarta Pathways - Descriptions}
\format{list containing 314 character values, the descriptions for the given
  pathways.}
\usage{
biocarta_descriptions
}
\description{
A list containing the descriptions for each human BioCarta pathway.
\emph{Generated on Oct 12, 2019.}
}
\keyword{datasets}
