% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biogrid_adj_list}
\alias{biogrid_adj_list}
\title{BioGRID PIN Adjacency List}
\format{
list containing 15231 vectors. Each vector is the set of gene symbols
of interactors B for each interactor A.
}
\usage{
biogrid_adj_list
}
\description{
An adjacency list of vectors containing interactors B for each interactor A
in the BioGRID protein-protein interaction network (The designations
"interactor A" and "interactor B" are arbitrary).
\emph{Generated on Oct 12, 2019.}
}
\keyword{datasets}
