% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_functions.R
\name{calculate_background_score}
\alias{calculate_background_score}
\title{Calculate Background Score}
\usage{
calculate_background_score(pin, number_of_iterations, scores_vec)
}
\arguments{
\item{pin}{\code{\link[igraph]{igraph}} object containing the protein-protein
interaction network}

\item{number_of_iterations}{the number of iterations}

\item{scores_vec}{vector of score for each gene in the PIN}
}
\value{
a list containing 2 elements \describe{
  \item{sampling_score_means}{vector of sampling score means}
  \item{sampling_score_stds}{vector of sampling score standard deviations}
}
}
\description{
Calculate Background Score
}
