% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_functions.R
\name{calculate_component_score}
\alias{calculate_component_score}
\title{Calculate Component Score}
\usage{
calculate_component_score(
  pin,
  scores_vec,
  compo,
  sampling_score_means,
  sampling_score_stds
)
}
\arguments{
\item{pin}{\code{\link[igraph]{igraph}} object containing the protein-protein
interaction network}

\item{scores_vec}{vector of score for each gene in the PIN}

\item{compo}{component (vector)}

\item{sampling_score_means}{vector of background sampling score means}

\item{sampling_score_stds}{vector of background sampling score standard deviations}
}
\value{
the score of the component
}
\description{
Calculate Component Score
}
