% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_markers_descriptions}
\alias{cell_markers_descriptions}
\title{Cell Markers - Descriptions}
\format{
named vector containing 496 character values, the descriptions for
the given human cell types.
}
\usage{
cell_markers_descriptions
}
\description{
A named vector containing descriptions of different cell types from different
tissues in human. Names of the vectors are Cell Ontology IDs (if available)
of the cell types in the following format: "tissue type, cancer type, cell name"
For more information, refer to the article:
Zhang X, Lan Y, Xu J, et al. CellMarker: a manually curated resource of cell
markers in human and mouse. Nucleic Acids Res. 2019;47(D1):D721-D728.
\emph{Generated on Jan 16, 2020.}
}
\keyword{datasets}
