% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_markers_gsets}
\alias{cell_markers_gsets}
\title{Cell Markers - Gene Sets}
\format{
list containing 496 vectors. Each vector corresponds to a cell marker
gene set for a given human cell type.
}
\usage{
cell_markers_gsets
}
\description{
A list containing the sets of genes that are cell markers of different cell
types from different tissues in human. Each element is a vector of cell marker
gene symbols for the given cell type. Names correspond to the Cell Ontology ID
(if available) of the cell type. For more information, refer to the article:
Zhang X, Lan Y, Xu J, et al. CellMarker: a manually curated resource of cell
markers in human and mouse. Nucleic Acids Res. 2019;47(D1):D721-D728.
\emph{Generated on Jan 16, 2020.}
}
\keyword{datasets}
