% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_java_version}
\alias{check_java_version}
\title{Check Java Version}
\usage{
check_java_version(version = NULL)
}
\arguments{
\item{version}{character vector containing the output of "java -version". If
NULL, result of \code{\link{fetch_java_version}} is used (default = NULL)}
}
\value{
only parses and checks whether the java version is >= 1.8
}
\description{
Check Java Version
}
\details{
this function was adapted from the CRAN package \code{sparklyr}
}
