% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{choose_clusters}
\alias{choose_clusters}
\title{Cluster Pathways and Dynamically Cut the Dendrogram}
\usage{
choose_clusters(result_df)
}
\arguments{
\item{result_df}{resulting data frame of the pathfindr main workflow.}
}
\value{
This function first calculates the pairwise distances between the
  pathways in the \code{result_df} data frame. Via a shiny HTML document, the
  hierarchical clustering dendrogram is visualized. In this HTML document,
  the user can select the value at which to cut the tree and the resulting
  representative pathways are presented as a table
}
\description{
Cluster Pathways and Dynamically Cut the Dendrogram
}
\examples{
choose_clusters(result_df)
}
