% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_pws.R
\name{cluster_pathways}
\alias{cluster_pathways}
\title{Calculate Pairwise Distances Between Given Pathways}
\usage{
cluster_pathways(pathway_ids, agg_method = "average", plot_heatmap = T,
  kegg_update = F)
}
\arguments{
\item{pathway_ids}{List that contains genes for selected pathways that are to
be clustered.}

\item{agg_method}{the agglomeration method to be used.}

\item{plot_heatmap}{boolean value indicating whether or not to plot the heat map
of pathway clustering}

\item{kegg_update}{boolean value indicating whether or not to update KEGG
pathway genes (default = FALSE)}
}
\value{
Pairwise distance matrix. See "Chen, Y. A. et al. Integrated pathway
  clusters with coherent biological themes for target prioritisation. PLoS
  One 9, e99030, doi:10.1371/journal.pone.0099030 (2014)." for details on the
  method of pathway clustering. Optionally plots a heat map of pathway
  clustering.
}
\description{
Calculate Pairwise Distances Between Given Pathways
}
\examples{
cluster_pathways(pathway_genes)
}
\seealso{
\code{\link[stats]{hclust}} for hierchical clustering,
  \code{\link[stats]{heatmap}} for drawing a heatmap.
}
