% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{configure_output_dir}
\alias{configure_output_dir}
\title{Configure Output Directory Name}
\usage{
configure_output_dir(output_dir = NULL)
}
\arguments{
\item{output_dir}{the directory to be created where the output and intermediate
files are saved (default = \code{NULL}, a temporary directory is used)}
}
\value{
/path/to/output/dir
}
\description{
Configure Output Directory Name
}
