% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_HTML_report}
\alias{create_HTML_report}
\title{Create HTML Report of pathfindR Results}
\usage{
create_HTML_report(input, input_processed, final_res, gene_sets)
}
\arguments{
\item{input}{the input data that pathfindR uses. The input must be a data
  frame with three columns: \enumerate{
  \item Gene Symbol (Gene Symbol)
  \item Change value, e.g. log(fold change) (OPTIONAL)
  \item p value, e.g. adjusted p value associated with differential expression
}}

\item{input_processed}{processed input data frame}

\item{final_res}{final pathfindR result data frame}

\item{gene_sets}{Name of the gene sets to be used for enrichment analysis.
Available gene sets are "KEGG", "Reactome", "BioCarta", "GO-All",
"GO-BP", "GO-CC", "GO-MF", "cell_markers", "mmu_KEGG" or "Custom".
If "Custom", the arguments \code{custom_genes} and \code{custom_descriptions}
must be specified. (Default = "KEGG")}
}
\description{
Create HTML Report of pathfindR Results
}
