% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{custom_result}
\alias{custom_result}
\title{Custom Gene Set Enrichment Results}
\format{
data frame containing 2 rows and 8 columns. Each row is a gene set (the TF target gene sets).
}
\usage{
custom_result
}
\description{
A data frame consisting of pathfindR enrichment analysis results on the
example TF target genes data (target gene sets of CREB and MYC).
\emph{Generated on Oct 22, 2019.}
}
\keyword{datasets}
