% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_score.R
\name{.component_new}
\alias{.component_new}
\title{Create a mutable component seeded with a single node}
\usage{
.component_new(sc, start_node)
}
\arguments{
\item{sc}{A score context.}

\item{start_node}{A single node name to start from.}
}
\value{
An environment representing the component.
}
\description{
Create a mutable component seeded with a single node
}
