% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_score.R
\name{.find_subnetworks}
\alias{.find_subnetworks}
\title{Find the scored subnetworks among a set of "on" nodes}
\usage{
.find_subnetworks(network, sc, on_names)
}
\arguments{
\item{network}{A network from \code{.build_network()}.}

\item{sc}{A score context.}

\item{on_names}{Character vector of node names that are switched on.}
}
\value{
A list of subnetwork objects.
}
\description{
Find the scored subnetworks among a set of "on" nodes
}
