% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_greedy.R
\name{.greedy_removal}
\alias{.greedy_removal}
\title{Greedy removal pass}
\usage{
.greedy_removal(gs, comp, removable)
}
\arguments{
\item{gs}{A greedy-state environment.}

\item{comp}{A mutable component environment.}

\item{removable}{An environment used as a set of removable nodes.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Tries to remove each removable node; a removal is accepted when it improves
the best score, otherwise the node is put back.
}
