% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_score.R
\name{.make_subnetwork}
\alias{.make_subnetwork}
\title{Create a subnetwork object from a vector of node names}
\usage{
.make_subnetwork(sc, nodes)
}
\arguments{
\item{sc}{A score context.}

\item{nodes}{Character vector of node names.}
}
\value{
A list with elements \code{nodes}, \code{zsum} and \code{score}.
}
\description{
Create a subnetwork object from a vector of node names
}
