% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_score.R
\name{.sort_subnetworks_desc}
\alias{.sort_subnetworks_desc}
\title{Sort a list of subnetwork objects by score, descending}
\usage{
.sort_subnetworks_desc(subs)
}
\arguments{
\item{subs}{A list of subnetwork objects.}
}
\value{
The list reordered so the highest scoring subnetwork is first.
}
\description{
Sort a list of subnetwork objects by score, descending
}
