% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{download_kegg_png}
\alias{download_kegg_png}
\title{Download Colored KEGG Diagram PNG}
\usage{
download_kegg_png(pw_url, f_path, quiet = TRUE)
}
\arguments{
\item{pw_url}{url to download}

\item{f_path}{local path to save the file}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and the
progress bar while downloading file(s)}
}
\value{
download status
}
\description{
Download Colored KEGG Diagram PNG
}
