% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{enrichment}
\alias{enrichment}
\title{Perform Enrichment Analysis for a Single Gene Set}
\usage{
enrichment(genes_by_pathway, genes_of_interest, pathways_list,
  adj_method = "bonferroni", enrichment_threshold, DEG_vec, all_genes)
}
\arguments{
\item{genes_by_pathway}{List that contains genes for each gene set. Names of
this list are gene set IDs.}

\item{genes_of_interest}{The set of gene symbols to be used for enrichment
analysis. In the scope of this package, these are genes that were
identified for an active subnetwork.}

\item{pathways_list}{List that contains pathway descriptions for the gene set
IDs. Names of this list are gene set IDs.}

\item{adj_method}{correction method to be used for adjusting p-values.}

\item{enrichment_threshold}{adjusted-p value threshold used when filtering
pathway enrichment results}

\item{DEG_vec}{vector of differentially-expressed gene symbols}

\item{all_genes}{vector of background genes}
}
\value{
A data frame that contains enrichment results.
}
\description{
Perform Enrichment Analysis for a Single Gene Set
}
\examples{
enrichment(kegg_genes, c("PER1", "PER2", "CRY1", "CREB1"), kegg_pathways,
           "bonferroni", 0.05, "PER1", unlist(kegg_genes))
}
\seealso{
\code{\link[stats]{p.adjust}} for adjustment of p values. See
  \code{\link{run_pathfindR}} for the wrapper function of the pathfindR
  workflow. \code{\link{hyperg_test}} for the details on hypergeometric
  distribution-based hypothesis testing.
}
