% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{enrichment_chart}
\alias{enrichment_chart}
\title{Plot the Bubble Chart of Enrichment Results}
\usage{
enrichment_chart(result_df)
}
\arguments{
\item{result_df}{a data frame consisting of at least 5 columns:\describe{
  \item{Pathway}{Description of the enriched pathway}
  \item{Fold_Enrichment}{Fold enrichment value for the enriched pathway}
  \item{lowest_p}{the lowest adjusted-p value of the given pathway over all iterations}
  \item{Up_regulated}{the up-regulated genes in the input involved in the given pathway, comma-separated}
  \item{Down_regulated}{the down-regulated genes in the input involved in the given pathway, comma-separated}
}}
}
\value{
a `ggplot2` object containing the bubble chart. The x-axis corresponds to
fold enrichment values while the y-axis indicates the enriched pathways. Size of
the bubble indicates the number of DEGs in the given pathway. Color indicates
the -log10(lowest-p) value. The closer the color is to red, the more significant
the enrichment is.
}
\description{
This function is used to plot a bubble chart displaying the enrichment
results.
}
\examples{
g <- enrichment_chart(RA_output)
}
