% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_active_snws}
\alias{example_active_snws}
\title{Example Active Subnetworks}
\format{list containing 166 vectors. Each vector is the set of genes for an active subnetwork.}
\usage{
example_active_snws
}
\description{
A list of vectors containing genes for each active subnetwork that passed the filtering step.
}
\keyword{datasets}
