% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata from a structured file path}
\usage{
extract_metadata(paths, regex, cols, validate)
}
\arguments{
\item{paths}{character vector of filenames.}

\item{regex}{a rege that matches files.}

\item{cols}{colunn names for groups captured in \code{regex}.}

\item{validate}{a named list of functions (names indicate cols) that validate metadata (e.g. is an integer, is date, etc.).}
}
\description{
This might be more internal; we can wrap this with friendly functions rather
than having users write a regex.
}

