% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{fetch_java_version}
\alias{fetch_java_version}
\title{Obtain Java Version}
\usage{
fetch_java_version()
}
\value{
character vector containing the output of 'java -version'
}
\description{
Obtain Java Version
}
\details{
this function was adapted from the CRAN package \code{sparklyr}
}
