% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_by_pathway}
\alias{genes_by_pathway}
\title{Genes by KEGG Pathway}
\format{list containing 325 vectors of gene symbols. Each vector corresponds
  to a pathway.}
\usage{
genes_by_pathway
}
\description{
A list containing the genes involved in each human KEGG pathway. Each
element is a vector of gene symbols located in the given pathway. Names
correspond to the KEGG ID of the pathway. This data was retrieved on Feb 26,
2018.
}
\keyword{datasets}
