% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{get_gene_sets_list}
\alias{get_gene_sets_list}
\title{Get Gene Sets List}
\usage{
get_gene_sets_list(source = "KEGG", ...)
}
\arguments{
\item{source}{As of this version, either "KEGG" or "Reactome" (default = "KEGG")}

\item{...}{Additional argument for \code{\link{get_kegg_gsets}}}
}
\value{
A list containing 2 elements: \itemize{
\item{gene_sets}{A list containing the genes involved in each gene set}
\item{descriptions}{A named vector containing the descriptions for each gene set}
}. For KEGG, it is possible to choose a specific organism. For a full list
of all available organisms, see \url{https://www.genome.jp/kegg/catalog/org_list.html}.
For Reactome, there is only one pathway gene sets collection.
}
\description{
Get Gene Sets List
}
