% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{get_kegg_gsets}
\alias{get_kegg_gsets}
\title{Retrieve Organism-specific KEGG Pathway Gene Sets}
\usage{
get_kegg_gsets(org_code = "hsa")
}
\arguments{
\item{org_code}{KEGG organism code for the selected organism. For a full list
of all available organisms, see \url{https://www.genome.jp/kegg/catalog/org_list.html}}
}
\value{
list containing 2 elements: \itemize{
\item{gene_sets}{A list containing the genes involved in each KEGG pathway}
\item{descriptions}{A named vector containing the descriptions for each KEGG pathway}
}
}
\description{
Retrieve Organism-specific KEGG Pathway Gene Sets
}
