% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{get_reactome_gsets}
\alias{get_reactome_gsets}
\title{Retrieve Reactome Pathway Gene Sets}
\usage{
get_reactome_gsets()
}
\value{
Gets the latest Reactome pathways gene sets in gmt format. Parses the
gmt file and returns a list containing 2 elements: \itemize{
\item{gene_sets}{A list containing the genes involved in each Reactome pathway}
\item{descriptions}{A named vector containing the descriptions for each Reactome pathway}
}
}
\description{
Retrieve Reactome Pathway Gene Sets
}
