% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_all_descriptions}
\alias{go_all_descriptions}
\title{Gene Ontology - All Ontology Descriptions}
\format{list containing 14586 character values, the descriptions for the given
go gene set.}
\usage{
go_all_descriptions
}
\description{
A list containing the descriptions for each human GO ontology term.
\emph{Generated on Sep 30, 2019.}
}
\keyword{datasets}
