% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_all_pathways}
\alias{go_all_pathways}
\title{Gene Ontology - All Ontology Descriptions}
\format{list containing 5123 character values, the descriptions for the given
pathways.}
\usage{
go_all_pathways
}
\description{
A list containing the descriptions for each human GO ontology category. This
data was retrieved on Nov 9, 2018.
}
\keyword{datasets}
