% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_bp_genes}
\alias{go_bp_genes}
\title{Gene Ontology - Biological Process Ontology Gene Sets}
\format{list containing 3941 vectors of gene symbols. Each vector corresponds
  to a gene set.}
\usage{
go_bp_genes
}
\description{
A list containing the genes involved in each GO Biological Process. Each
element is a vector of gene symbols located in the given gene set. This data
was retrieved on May 13, 2018.
}
\keyword{datasets}
