% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_bp_pathways}
\alias{go_bp_pathways}
\title{Gene Ontology - Biological Process Ontology Descriptions}
\format{list containing 3941 character values, the descriptions for the given
  pathways.}
\usage{
go_bp_pathways
}
\description{
A list containing the descriptions for each human GO Biological Process. This
data was retrieved on May 13, 2018.
}
\keyword{datasets}
