% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_cc_pathways}
\alias{go_cc_pathways}
\title{Gene Ontology - Cellular Component Ontology Descriptions}
\format{list containing 470 character values, the descriptions for the given
  pathways.}
\usage{
go_cc_pathways
}
\description{
A list containing the descriptions for each human GO Cellular Component. This
data was retrieved on May 13, 2018.
}
\keyword{datasets}
