% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_mf_pathways}
\alias{go_mf_pathways}
\title{Gene Ontology - Molecular Function Ontology Descriptions}
\format{list containing 713 character values, the descriptions for the given
  pathways.}
\usage{
go_mf_pathways
}
\description{
A list containing the descriptions for each human GO Molecular Function. This
data was retrieved on May 13, 2018.
}
\keyword{datasets}
