% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_functions.R
\name{greedy_breadth_first_active_subnetwork_search}
\alias{greedy_breadth_first_active_subnetwork_search}
\title{Greedy Breadth-First Active Subnetwork search}
\usage{
greedy_breadth_first_active_subnetwork_search(
  seed_node,
  pin,
  scores_df,
  scores_vec,
  sampling_result,
  max_depth,
  check_second_neighbors
)
}
\arguments{
\item{seed_node}{the seed node (an igraph.vs object)}

\item{pin}{\code{\link[igraph]{igraph}} object containing the protein-protein
interaction network}

\item{scores_df}{data frame containing scores per each gene in the PIN}

\item{scores_vec}{vector of score for each gene in the PIN}

\item{sampling_result}{list containing vector of bacground sampling score means
and vector of background sampling score standard deviations}

\item{max_depth}{maximum depth in search}

\item{check_second_neighbors}{boolean to indicate whether to check second
neighbors when any direct neighbor does not improve the score}
}
\value{
an active module (vector)
}
\description{
Greedy Breadth-First Active Subnetwork search
}
