% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{gset_list_from_gmt}
\alias{gset_list_from_gmt}
\title{Retrieve Gene Sets from GMT-format File}
\usage{
gset_list_from_gmt(path2gmt, descriptions_idx = 2)
}
\arguments{
\item{path2gmt}{path to the gmt file}

\item{descriptions_idx}{index for descriptions (default = 2)}
}
\value{
list containing 2 elements: \itemize{
\item{gene_sets - A list containing the genes involved in each gene set}
\item{descriptions - A named vector containing the descriptions for each gene set}
}
}
\description{
Retrieve Gene Sets from GMT-format File
}
