% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intact_adj_list}
\alias{intact_adj_list}
\title{IntAct PIN Adjacency List}
\format{
list containing 15057 vectors. Each vector is the set of gene symbols
of interactors B for each interactor A.
}
\usage{
intact_adj_list
}
\description{
An adjacency list of vectors containing interactors B for each interactor A
in the IntAct protein-protein interaction network (The designations
"interactor A" and "interactor B" are arbitrary).
\emph{Generated on Oct 19, 2019.}
}
\keyword{datasets}
