% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{isColor}
\alias{isColor}
\title{Check if value is a valid color}
\usage{
isColor(x)
}
\arguments{
\item{x}{value}
}
\value{
TRUE if x is a valid color, otherwise FALSE
}
\description{
Check if value is a valid color
}
