% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-visualization.R
\name{isOrderable}
\alias{isOrderable}
\title{Check if column can be ordered}
\usage{
isOrderable(order_by, result_df)
}
\arguments{
\item{order_by}{A column name in `result_df`}

\item{result_df}{A \link[base]{data.frame}}
}
\value{
A \link[base]{data.frame} or an error message
}
\description{
Check if column can be ordered
}
