% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_adj_list}
\alias{kegg_adj_list}
\title{KEGG PIN Adjacency List}
\format{
list containing 4507 vectors. Each vector is the set of gene symbols
of interactors B for each interactor A.
}
\usage{
kegg_adj_list
}
\description{
An adjacency list of vectors containing interactors B for each interactor A
in the KEGG protein-protein interaction network (The designations
"interactor A" and "interactor B" are arbitrary).
\emph{Generated on Oct 12, 2019.}
}
\keyword{datasets}
