% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_pathways}
\alias{kegg_pathways}
\title{KEGG Pathway Descriptions}
\format{vector containing 321 character values, the descriptions for the given
  pathways.}
\usage{
kegg_pathways
}
\description{
A list containing the descriptions for each human KEGG pathway. Names of the
list correspond to the KEGG ID of the pathway. Pathways that did not contain
any genes were discarded. This data was retrieved on Nov 9, 2018.
}
\keyword{datasets}
