% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu_kegg_descriptions}
\alias{mmu_kegg_descriptions}
\title{Mus Musculus KEGG Pathways - Descriptions}
\format{
named vector containing 323 character values, the descriptions for the given
  pathways.
}
\usage{
mmu_kegg_descriptions
}
\description{
A named vector containing the descriptions for each Mus musculus KEGG pathway. Names of the
vector correspond to the KEGG ID of the pathway. Pathways that did not contain
any genes were discarded.
\emph{Generated on Oct 28, 2019.}
}
\keyword{datasets}
