% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu_kegg_genes}
\alias{mmu_kegg_genes}
\title{Mus Musculus KEGG Pathways - Gene Sets}
\format{list containing 323 vectors of gene symbols. Each vector corresponds
  to a pathway.}
\usage{
mmu_kegg_genes
}
\description{
A list containing the genes involved in each Mus musculus KEGG pathway. Each element
is a vector of gene symbols located in the given pathway. Names correspond to
the KEGG ID of the pathway. Pathways that did not contain any genes were
discarded.
\emph{Generated on Oct 28, 2019.}
}
\keyword{datasets}
