% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{obtain_colored_url}
\alias{obtain_colored_url}
\title{Obtain URL for a KEGG pathway diagram with a given set of genes marked}
\usage{
obtain_colored_url(pw_id, KEGG_gene_ids, fg_cols, bg_cols)
}
\arguments{
\item{pw_id}{KEGG pathway ID}

\item{KEGG_gene_ids}{KEGG gene IDs for marking}

\item{fg_cols}{colors for the text and border}

\item{bg_cols}{background colors of the objects in a pathway diagram.}
}
\value{
download status
}
\description{
Obtain URL for a KEGG pathway diagram with a given set of genes marked
}
