% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{obtain_KEGGML_URL}
\alias{obtain_KEGGML_URL}
\title{Obtain KGML file for a KEGG pathway (hsa)}
\usage{
obtain_KEGGML_URL(pw_id, pwKGML, quiet = TRUE)
}
\arguments{
\item{pw_id}{KEGG pathway ID}

\item{pwKGML}{destination file}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and the
progress bar while downloading file(s)}
}
\value{
KGML URL
}
\description{
Obtain KGML file for a KEGG pathway (hsa)
}
