% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseActive.R
\name{parseActiveSnwSearch}
\alias{parseActiveSnwSearch}
\title{Parse Active Subnetwork Search Output File}
\usage{
parseActiveSnwSearch(output_path, signif_genes)
}
\arguments{
\item{output_path}{path to the output of an Active Subnetwork Search.}

\item{signif_genes}{the vector of significant genes.}
}
\value{
A list of genes in every active subnetwork that has a score > 3 and
  that has at least 2 significant genes.
}
\description{
Parse Active Subnetwork Search Output File
}
\examples{
\dontrun{
parseActiveSnwSearch(output, significant_genes)
}
}
\seealso{
See \code{\link{run_pathfindr}} for the wrapper function of the
  pathfindr workflow
}
