% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsejActive.R
\name{parsejActive}
\alias{parsejActive}
\title{Parse jActive Output File}
\usage{
parsejActive(jactive_output, signif_genes)
}
\arguments{
\item{jactive_output}{the output of a jActive run.}

\item{signif_genes}{the vector of significant genes.}
}
\value{
A list of genes in every active subnetwork that has a score > 0 and
  that has at least 2 significant genes.
}
\description{
Parse jActive Output File
}
\examples{
parsejActive(output, significant_genes)
}
\seealso{
See \code{\link{run_pathfindr}} for the wrapper function of the
  pathfindr workflow
}
