\name{pathfindr-package}
\alias{pathfindr-package}
\alias{pathfindr}
\docType{package}
\title{
\packageTitle{pathfindr}
}
\description{
\packageDescription{pathfindr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pathfindr}
\packageIndices{pathfindr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pathfindr}

Maintainer: \packageMaintainer{pathfindr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
