% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{pathfindr}
\alias{pathfindr}
\title{Wrapper Function for pathfindr}
\usage{
pathfindr(input, p_val_threshold = 0.05, enrichment_threshold = 1e-04,
  adj_method = "bonferroni", iterations = 10, n_threads = NULL,
  n_snw = 1000, overlap_threshold = 0.5, kegg_update = F)
}
\arguments{
\item{input}{the input data that pathfindr uses. The input must be a data
frame with three columns: \enumerate{ \item Gene Symbol (HGNC Gene
Symbol) \item Change value, e.g. log(fold change) \item adjusted p value
associated with test, e.g. differential expression/methylation }}

\item{p_val_threshold}{the adjusted-p value threshold to use when filtering
the input data frame. Must a numeric value between 0 and 1.}

\item{enrichment_threshold}{threshold used when filtering individual pathway
enrichment results}

\item{adj_method}{correction method to be used for adjusting p-values of
pathway enrichment results}

\item{iterations}{number of iterations for active subnetwork search and
enrichment analyses}

\item{n_threads}{optional argument for specifying the number of threads. The
function determines this automatically}

\item{n_snw}{optional argument for specifying the number of active
subnetworks when running jActive modules. (Default = 1000)}

\item{overlap_threshold}{optional argument for specifying the overlap
thresholds when running jActive modules. (Default = 0.5)}

\item{kegg_update}{boolean value indicating whether or not to update KEGG
pathway genes (default = FALSE)}
}
\value{
Data frame of pathview enrichment results. Columns are: "ID",
  "Pathway", "occurence", "lowest_p", "highest_p". "ID" is the KEGG ID for a
  given pathway and "Pathway" is the name. "occurence" indicates the number
  of times the given pathway is encountered over all iterations. "lowest_p"
  and "highest_p" indicate the lowest and highest adjusted p values over all
  iterations.
}
\description{
Wrapper Function for pathfindr
}
\examples{
pathfindr(input_data_frame)
}
\seealso{
\code{\link{input_testing}} for input testing,
  \code{\link{input_processing}} for input processing,
  \code{\link{current_KEGG}} for KEGG pathway genes retrieval,
  \code{\link{parsejActive}} for parsing a jActive modules output,
  \code{\link{enrichment}} for pathway enrichment analysis and
  \code{\link{pathmap}} for annotation of involved genes and visualization of
  pathways.
}
