% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathmap.R
\name{pathmap}
\alias{pathmap}
\title{Annotate Involved Genes In Pathways and Visualize Pathways}
\usage{
pathmap(pw_table, gene_data)
}
\arguments{
\item{pw_table}{Data frame of enrichment results. Columns are: "ID",
"Pathway", "occurence", "lowest_p", "highest_p".}

\item{gene_data}{Single column data frame containing change values (e.g.
log(fold change) values) for significant genes.}
}
\value{
Data frame of enrichment results with genes involved in each pathway
  presented. Columns are: "ID", "Pathway", "occurence", "lowest_p",
  "highest_p","Involved_genes". The function also creates visualizations of
  the pathways with the package "pathview".
}
\description{
Annotate Involved Genes In Pathways and Visualize Pathways
}
\examples{
pathmap(pathway_table, change_data)
}
\seealso{
\code{\link[pathview]{pathview}} for pathway based data integration
  and visualization. See \code{\link{run_pathfindr}} for the wrapper function
  of the pathfindr workflow
}
