% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways_list}
\alias{pathways_list}
\title{KEGG Pathway Descriptions}
\format{list containing 325 character values, the descriptions for the given
  pathways.}
\usage{
pathways_list
}
\description{
A list containing the descriptions for each human KEGG pathway. Names of the
list correspond to the KEGG ID of the pathway. This data was retrieved on Feb
26, 2018.
}
\keyword{datasets}
