% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{process_pin}
\alias{process_pin}
\title{Process Data frame of Protein-protein Interactions}
\usage{
process_pin(pin_df)
}
\arguments{
\item{pin_df}{data frame of protein-protein interactions with 2 columns:
"Interactor_A" and "Interactor_B"}
}
\value{
processed PIN data frame (removes self-interactions and
duplicated interactions)
}
\description{
Process Data frame of Protein-protein Interactions
}
\examples{
toy_pin <- data.frame(Interactor_A = c(LETTERS[1:4], LETTERS[5:8], "Z"),
                      Interactor_B = c(LETTERS[5:8], LETTERS[1:4], "Z"),
                      stringsAsFactors = FALSE)
processed_pin <- pathfindR:::process_pin(toy_pin)
}
