% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RA_exp_mat}
\alias{RA_exp_mat}
\title{Example Input for pathfindR - pathway z-scores}
\format{A matrix with 572 rows and 33 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE15573}
}
\usage{
RA_exp_mat
}
\description{
A matrix containing the log2-normalized expression values of the differentially-expressed genes
for 18 rheumatoid arthritis (RA) patients and 15 healthy subjects. Expression values of
differentially-expressed genes with adj.P.Val <= 0.05 are presented in this dataset.
}
\keyword{datasets}
