% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RA_input}
\alias{RA_input}
\title{Example Input for the pathfindr Enrichment Workflow - Rheumatoid Arthritis}
\format{A data frame with 857 rows and 3 variables:
\describe{
  \item{Gene.symbol}{HGNC gene symbols of the differentially-expressed genes}
  \item{logFC}{log2-fold-change values}
  \item{adj.P.Val}{adjusted p values, via the Benjamini & Hochberg(1995) method}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE15573}
}
\usage{
RA_input
}
\description{
A dataset containing differentially-expressed genes along with the associated
log2-fold-changes and adjusted p-values for the GEO data set GSE15573. This dataset
aimed to characterize gene expression profiles in the peripheral blood mononuclear cells
of rheumatoid arthritis (RA) patients versus healthy subjects. Differentially-expressed
genes with |logFC| >= 1 and adj.P.Val <= 0.05 are presented in this dataset.
}
\seealso{
\code{\link{RA_output}} for example output.
}
\keyword{datasets}
