% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RA_output}
\alias{RA_output}
\title{Example Output for the pathfindR Enrichment Workflow - Rheumatoid Arthritis}
\format{A data frame with 21 rows and 7 columns:
\describe{
  \item{ID}{KEGG ID of the enriched pathway}
  \item{Pathway}{Description of the enriched pathway}
  \item{occurence}{the number of iterations that the given pathway was found to enriched over all iterations}
  \item{lowest_p}{the lowest adjusted-p value of the given pathway over all iterations}
  \item{highest_p}{the highest adjusted-p value of the given pathway over all iterations}
  \item{Up_regulated}{the upregulated genes in the input involved in the given pathway, comma-separated}
  \item{Down_regulated}{the downregulated genes in the input involved in the given pathway, comma-separated}
}}
\usage{
RA_output
}
\description{
A dataset containing the results of pathfindR's active subnetwork search and
pathway enrichment workflow performed on the rheumatoid arthritis
differential-expression dataset \code{RA_input}.
}
\seealso{
\code{\link{RA_input}} for example input.
}
\keyword{datasets}
