% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_descriptions}
\alias{reactome_descriptions}
\title{Reactome Pathways - Descriptions}
\format{list containing 2263 character values, the descriptions for the given
  pathways.}
\usage{
reactome_descriptions
}
\description{
A list containing the descriptions for each human Reactome pathway. Names of the
list correspond to the Reactome ID of the pathway.
\emph{Generated on Oct 12, 2019.}
}
\keyword{datasets}
