% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome_pathways}
\alias{reactome_pathways}
\title{Reactome Pathway Descriptions}
\format{list containing 2022 character values, the descriptions for the given
  pathways.}
\usage{
reactome_pathways
}
\description{
A list containing the descriptions for each human Reactome pathway. Names of the
list correspond to the Reactome ID of the pathway. This data was retrieved on May
13, 2018.
}
\keyword{datasets}
