% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_path.R
\name{return_pin_path}
\alias{return_pin_path}
\title{Return Path to Given Protein Interaction Network (PIN)}
\usage{
return_pin_path(pin_name = "GeneMania")
}
\arguments{
\item{pin_name}{Name of the chosen PIN. Must be one of c("Biogrid", "STRING",
"GeneMania", "BioPlex"). Defaults to "GeneMania".}
}
\value{
A character value that contains the path to chosen PIN.
}
\description{
Return Path to Given Protein Interaction Network (PIN)
}
\examples{
pin_path <- return_pin_path("Biogrid")
}
