% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_pathfindr}
\alias{run_pathfindr}
\title{Wrapper Function for pathfindr Workflow}
\usage{
run_pathfindr(input, p_val_threshold = 0.05, enrichment_threshold = 1e-04,
  adj_method = "bonferroni", iterations = 10, n_processes = NULL,
  n_snw = 1000, overlap_threshold = 0.5, kegg_update = F,
  pin_name = "IntAct")
}
\arguments{
\item{input}{the input data that pathfindr uses. The input must be a data
frame with three columns: \enumerate{ \item Gene Symbol (HGNC Gene
Symbol) \item Change value, e.g. log(fold change) \item adjusted p value
associated with test, e.g. differential expression/methylation }}

\item{p_val_threshold}{the adjusted-p value threshold to use when filtering
the input data frame. Must a numeric value between 0 and 1.}

\item{enrichment_threshold}{threshold used when filtering individual pathway
enrichment results}

\item{adj_method}{correction method to be used for adjusting p-values of
pathway enrichment results}

\item{iterations}{number of iterations for active subnetwork search and
enrichment analyses}

\item{n_processes}{optional argument for specifying the number of processes
used by foreach. The function determines this automatically}

\item{n_snw}{optional argument for specifying the number of active
subnetworks when running jActive modules. (Default = 1000)}

\item{overlap_threshold}{optional argument for specifying the overlap
thresholds when running jActive modules. (Default = 0.5)}

\item{kegg_update}{boolean value indicating whether or not to update KEGG
pathway genes (default = FALSE)}

\item{pin_name}{Name of the chosen PIN. Must be one of c("Biogrid", "STRING",
"GeneMania", "BioPlex", "HitPredict", "IntAct"). Defaults to "IntAct".}
}
\value{
Data frame of pathview enrichment results. Columns are: "ID",
  "Pathway", "occurence", "lowest_p", "highest_p". "ID" is the KEGG ID for a
  given pathway and "Pathway" is the name. "occurence" indicates the number
  of times the given pathway is encountered over all iterations. "lowest_p"
  and "highest_p" indicate the lowest and highest adjusted p values over all
  iterations. The function also creates an HTML report with the pathview
  enrichment results linked to the visualizations of the pathways in addition
  to the table of converted gene symbols. This report can be found in
  "results.html".
}
\description{
\code{run_pathfindr} is the wrapper function for the pathfindr workflow
}
\details{
This function takes in a data frame consisting of Gene Symbol,
log-fold-change and adjusted-p values. After input testing, any gene symbols
that are not in the PIN are converted to alias symbols if the alias is in the
PIN. Next, active subnetwork search is performed. Pathway enrichment analysis
is performed using the genes in each of the active subnetworks. Pathways with
adjusted-p values lower than \code{enrichment_threshold} are discarded. The
lowest adjusted-p value (over all subnetworks) for each pathway is kept. This
process of active subnetwork search and enrichment is repeated  for a
selected number of \code{iterations}, which is done in parallel. Over all
iterations, the lowest and the highest adjusted-p values, as well as number
of occurences are reported for each enriched pathway.
}
\section{Warning}{
 Depending on the protein interaction network of your
  choice, active subnetwork finding component of pathfindr may take a very
  long time to finish. Therefore, overnight runs are recommended.
}

\examples{
run_pathfindr(input_data_frame)
}
\seealso{
\code{\link{input_testing}} for input testing,
  \code{\link{input_processing}} for input processing,
  \code{\link{current_KEGG}} for KEGG pathway genes retrieval,
  \code{\link{parsejActive}} for parsing a jActive modules output,
  \code{\link{enrichment}} for pathway enrichment analysis and
  \code{\link{pathmap}} for annotation of involved genes and visualization of
  pathways. See \code{\link[foreach]{foreach}} for details on parallel
  execution of looping constructs. See \code{\link{choose_clusters}} for
  clustering the resulting enriched pathways.
}
