% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{score_terms}
\alias{score_terms}
\title{Calculate Agglomerated Scores of Enriched Terms for Each Subject}
\usage{
score_terms(enrichment_table, exp_mat, cases = NULL,
  use_description = FALSE, plot_hmap = TRUE, ...)
}
\arguments{
\item{enrichment_table}{a data frame that must contain the 3 columns below: \describe{
  \item{Term_Description}{Description of the enriched term (necessary if \code{use_description = TRUE})}
  \item{ID}{ID of the enriched term (necessary if \code{use_description = FALSE})}
  \item{Up_regulated}{the up-regulated genes in the input involved in the given term's gene set, comma-separated}
  \item{Down_regulated}{the down-regulated genes in the input involved in the given term's gene set, comma-separated}
}}

\item{exp_mat}{the experiment (e.g., gene expression/methylation) matrix.
Columns are samples and rows are genes. Column names must contain sample
names and row names must contain the gene symbols.}

\item{cases}{(Optional) A vector of sample names that are cases in the
case/control experiment. (default = NULL)}

\item{use_description}{Boolean argument to indicate whether term descriptions
(in the "Term_Description" column) should be used. (default = \code{FALSE})}

\item{plot_hmap}{Boolean value to indicate whether or not to draw the
heatmap plot of the scores. (default = TRUE)}

\item{...}{Additional arguments for \code{\link{plot_scores}} for aesthetics
of the heatmap plot}
}
\value{
Matrix of agglomerated scores of each enriched term per sample.
Columns are samples, rows are enriched terms. Optionally, displays a heatmap
of this matrix.
}
\description{
Calculate Agglomerated Scores of Enriched Terms for Each Subject
}
\examples{
score_matrix <- score_terms(RA_output, RA_exp_mat, plot_hmap = FALSE)
}
