% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{string_adj_list}
\alias{string_adj_list}
\title{STRING PIN Adjacency List}
\format{
list containing 11934 vectors. Each vector is the set of gene symbols
of interactors B for each interactor A.
}
\usage{
string_adj_list
}
\description{
An adjacency list of vectors containing interactors B for each interactor A
in the STRING protein-protein interaction network (The designations
"interactor A" and "interactor B" are arbitrary). Only interactions with a combined
score >= 800 were kept.
\emph{Generated on Oct 31, 2019.}
}
\keyword{datasets}
