% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biocarta_genes}
\alias{biocarta_genes}
\title{BioCarta Gene Sets}
\format{list containing 217 vectors of gene symbols. Each vector corresponds
  to a gene set.}
\usage{
biocarta_genes
}
\description{
A list containing the genes involved in each human BioCarta pathway. Each
element is a vector of gene symbols located in the given pathway. This data
was retrieved on May 13, 2018.
}
\keyword{datasets}
