% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biocarta_pathways}
\alias{biocarta_pathways}
\title{BioCarta Pathway Descriptions}
\format{list containing 217 character values, the descriptions for the given
  pathways.}
\usage{
biocarta_pathways
}
\description{
A list containing the descriptions for each human Reactome pathway. This data
was retrieved on May 13, 2018.
}
\keyword{datasets}
