% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_pathways}
\alias{cluster_pathways}
\title{Cluster Pathways}
\usage{
cluster_pathways(enrichment_res, method = "hierarchical",
  plot_clusters_graph = TRUE, use_names = FALSE, ...)
}
\arguments{
\item{enrichment_res}{data frame of pathway enrichment results
(result of `run_pathfindR`)}

\item{method}{Either "hierarchical" or "fuzzy". Details of clustering are
provided in the corresponding functions.}

\item{plot_clusters_graph}{boolean value indicate whether or not to plot
the graph diagram of clustering results (default = TRUE)}

\item{use_names}{boolean to indicate whether to use pathway names instead of
IDs (default = FALSE, i.e. use IDs)}

\item{...}{additional arguments for `create_kappa_matrix`,
`hierarchical_pw_clustering`, `fuzzy_pw_clustering` and `cluster_graph_vis`.
See documentation of these functions for more details.}
}
\value{
a data frame of clustering results. For "hierarchical", the cluster
assignments (Cluster) and whether the term is representative of its cluster
(Status) is added as columns. For "fuzzy", terms that are in multiple
clusters are provided for each cluster. The cluster assignments (Cluster)
and whether the term is representative of its cluster (Status) is
added as columns.
}
\description{
Cluster Pathways
}
\examples{
example_clustered <- cluster_pathways(RA_output[1:3,],
plot_clusters_graph = FALSE)
example_clustered <- cluster_pathways(RA_output[1:3,],
method = "fuzzy", plot_clusters_graph = FALSE)

}
\seealso{
See \code{\link{hierarchical_pw_clustering}} for hierarchical
clustering of enriched terms.
See \code{\link{fuzzy_pw_clustering}} for fuzzy clustering of enriched terms.
See \code{\link{cluster_graph_vis}} for graph visualization of clustering.
}
