% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_HTML_report}
\alias{create_HTML_report}
\title{Create HTML Report of pathfindR Results}
\usage{
create_HTML_report(input, input_processed, final_res)
}
\arguments{
\item{input}{the input data that pathfindR uses. The input must be a data
  frame with three columns: \enumerate{
  \item Gene Symbol (Gene Symbol)
  \item Change value, e.g. log(fold change) (OPTIONAL)
  \item p value, e.g. adjusted p value associated with differential expression
}}

\item{input_processed}{processed input data frame}

\item{final_res}{final pathfindR result data frame}
}
\description{
Create HTML Report of pathfindR Results
}
