% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_KEGG.R
\name{current_KEGG}
\alias{current_KEGG}
\title{Get Current KEGG Pathway Genes}
\usage{
current_KEGG()
}
\value{
list containing genes for each pathway. The names of the list are
  KEGG IDs.
}
\description{
Get Current KEGG Pathway Genes
}
\examples{
current_KEGG()
}
\seealso{
\code{\link{enrichment}} for pathway enrichment analysis.
}
