% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_score.R
\name{.find_components_named}
\alias{.find_components_named}
\title{Find the connected components among a set of "on" nodes}
\usage{
.find_components_named(network, on_names)
}
\arguments{
\item{network}{A network from \code{build_network()}.}

\item{on_names}{Character vector of node names that are switched on.}
}
\value{
A list of character vectors, one per component.
}
\description{
Returns one group of node names per connected component of the subgraph
induced by \code{on_names} (isolated nodes form singleton components).
}
