% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_greedy.R
\name{.greedy_filter}
\alias{.greedy_filter}
\title{Filter subnetworks by mutual overlap}
\usage{
.greedy_filter(comps, overlap_threshold, subnetwork_num)
}
\arguments{
\item{comps}{Score-sorted list of subnetwork objects (desc).}

\item{overlap_threshold}{Numeric threshold in (0, 1).}

\item{subnetwork_num}{Maximum number of subnetworks to return.}
}
\value{
Filtered list.
}
\description{
Filter subnetworks by mutual overlap
}
