% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_greedy.R
\name{.greedy_removal_idx}
\alias{.greedy_removal_idx}
\title{Greedy removal pass}
\usage{
.greedy_removal_idx(vstate, z_vec, sc_means, sc_stds, size, zsum, best_score)
}
\arguments{
\item{vstate}{Environment with mutable node vectors.}

\item{z_vec}{Numeric z-score vector.}

\item{sc_means}{MC means.}

\item{sc_stds}{MC stds.}

\item{size}{Current component size (from expand return value).}

\item{zsum}{Current z-score sum (from expand return value).}

\item{best_score}{Best score so far.}
}
\value{
Updated best_score.
}
\description{
Greedy removal pass
}
