% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_greedy.R
\name{.greedy_removal_idx}
\alias{.greedy_removal_idx}
\title{Greedy removal pass using vector indexing}
\usage{
.greedy_removal_idx(
  comp_state,
  comp_members,
  removable_vec,
  node2predecessor,
  node2dependent_count,
  sc
)
}
\value{
Invisibly \code{NULL}.
}
\description{
Greedy removal pass using vector indexing
}
