% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_greedy.R
\name{.greedy_search}
\alias{.greedy_search}
\title{Run the greedy active subnetwork search}
\usage{
.greedy_search(network, score_context, params, verbose = FALSE)
}
\arguments{
\item{network}{A network from \code{build_network()}.}

\item{score_context}{A score context from \code{build_score_context()}.}

\item{params}{A list of run parameters.}

\item{verbose}{Logical; emit progress messages.}
}
\value{
A list of subnetwork objects.
}
\description{
Run the greedy active subnetwork search
}
