% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Perform Enrichment Analysis}
\usage{
enrichment(genes_by_pathway, genes_of_interest, pathways_list,
  adj_method = "bonferroni")
}
\arguments{
\item{genes_by_pathway}{List that contains genes for each pathway. Names of
this list are the KEGG IDs.}

\item{genes_of_interest}{The set of gene symbols to be used for enrichment
analysis. In the scope of this package, these are genes that were
identified for a subnetwork.}

\item{pathways_list}{List that contains pathway descriptions for KEGG pathway
IDs. Names of this list are the KEGG IDs.}

\item{adj_method}{correction method to be used for adjusting p-values.}
}
\value{
A data frame that contains enrichment results.
}
\description{
Perform Enrichment Analysis
}
\examples{
enrichment(genes_by_pathway, genes_of_interest, pathways_list)
}
\seealso{
\code{\link[stats]{p.adjust}} for adjustment of p values.
}
