% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{enrichment_analyses}
\alias{enrichment_analyses}
\title{Perform Enrichment Analyses on the Input Subnetworks}
\usage{
enrichment_analyses(snws, sig_genes_vec, pin_path, genes_by_term,
  term_descriptions, adj_method = "bonferroni",
  enrichment_threshold = 0.05, list_active_snw_genes = FALSE)
}
\arguments{
\item{snws}{a list of subnetwork genes (i.e., vectors of genes for each subnetwork)}

\item{sig_genes_vec}{vector of significant gene symbols. In the scope of this
package, these are the input genes that were used for active subnetwork search}

\item{pin_path}{path to the Protein Interaction Network (PIN) file used in
the analysis}

\item{genes_by_term}{List that contains genes for each gene set. Names of
this list are gene set IDs}

\item{term_descriptions}{Vector that contains term descriptions for the
gene sets. Names of this vector are gene set IDs}

\item{adj_method}{correction method to be used for adjusting p-values of}

\item{enrichment_threshold}{adjusted-p value threshold used when filtering
enrichment results (default = 0.05)}

\item{list_active_snw_genes}{boolean value indicating whether or not to report
the non-DEG active subnetwork genes for the active subnetwork which was enriched for
the given term with the lowest p value (default = \code{FALSE})}
}
\value{
a dataframe of combined enrichment results. Columns are: \describe{
  \item{ID}{ID of the enriched term}
  \item{Term_Description}{Description of the enriched term}
  \item{Fold_Enrichment}{Fold enrichment value for the enriched term}
  \item{p_value}{p value of enrichment}
  \item{adj_p}{adjusted p value of enrichment}
  \item{non_DEG_Active_Snw_Genes (OPTIONAL)}{the non-DEG active subnetwork genes, comma-separated}
}
}
\description{
Perform Enrichment Analyses on the Input Subnetworks
}
\examples{
pin_path <- return_pin_path()
enr_res <- enrichment_analyses(snws = example_active_snws[1:2],
                               sig_genes_vec = RA_input$Gene.symbol[1:25],
                               pin_path = pin_path
                               genes_by_term = kegg_genes,
                               term_descriptions = kegg_descriptions)
}
\seealso{
\code{\link{enrichment}} for the enrichment analysis for a single gene set
}
