% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_unfiltered_snws}
\alias{example_unfiltered_snws}
\title{Example Unfiltered Active Subnetworks}
\format{
A list with 1000 elements. Each containing a named list:
\describe{
  \item{nodes}{active subnetwork nodes}
  \item{score}{score of the active subnetwork}
}
}
\usage{
example_unfiltered_snws
}
\description{
List containing unfiltered active subnetworks returned by
\code{\link{active_subnetwork_search}}
}
\keyword{datasets}
