% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_identification_pipeline.R
\name{get_active_subnetworks}
\alias{get_active_subnetworks}
\title{Get Active Subnetworks}
\usage{
get_active_subnetworks(
  input_for_search,
  pin_name_path = "Biogrid",
  score_quan_thr = 0.8,
  sig_gene_thr = 0.02,
  search_method = "GR",
  seed_for_stochastic_methods = 1234,
  verbose = FALSE,
  start_with_all_positives = FALSE,
  gene_init_prob = 0.1,
  sa_initial_temp = 1,
  sa_final_temp = 0.01,
  sa_iterations = 10000,
  ga_population_size = 400,
  ga_iterations = 10000,
  ga_crossover_rate = 1,
  ga_mutation_rate = 0,
  gr_max_depth = 1,
  gr_search_depth = 1,
  gr_overlap_threshold = 0.5,
  gr_subnetwork_num = 1000
)
}
\arguments{
\item{input_for_search}{input the input data that active subnetwork search uses. The input
must be a data frame containing at least these 2 columns: \describe{
  \item{GENE}{Gene Symbol}
  \item{P_VALUE}{p value obtained through a test, e.g. differential expression/methylation}
}}

\item{pin_name_path}{Name of the chosen PIN or absolute/path/to/PIN.sif. If PIN name,
must be one of c('Biogrid', 'STRING', 'GeneMania', 'IntAct', 'KEGG', 'mmu_STRING'). If
path/to/PIN.sif, the file must comply with the PIN specifications. (Default = 'Biogrid')}

\item{score_quan_thr}{active subnetwork score quantile threshold. Must be
between 0 and 1 or set to -1 for not filtering. (Default = 0.8)}

\item{sig_gene_thr}{threshold for the minimum proportion of significant genes in
the subnetwork (Default = 0.02) If the number of genes to use as threshold is
calculated to be < 2 (e.g. 50 signif. genes x 0.01 = 0.5), the threshold number
is set to 2}

\item{search_method}{algorithm to use when performing active subnetwork
search. Options are greedy search (GR), simulated annealing (SA) or genetic
algorithm (GA) for the search (default = 'GR').}

\item{seed_for_stochastic_methods}{seed for reproducibility while running active subnetwork search (applies for GR and SA)}

\item{verbose}{boolean value indicating whether to print messages (default=FALSE)}

\item{start_with_all_positives}{if TRUE: in GA, adds an individual with all positive
nodes. In SA, initializes candidate solution with all positive nodes. (default = FALSE)}

\item{gene_init_prob}{For SA and GA, probability of adding a gene in initial solution (default = 0.1)}

\item{sa_initial_temp}{Initial temperature for SA (default = 1.0)}

\item{sa_final_temp}{Final temperature for SA (default = 0.01)}

\item{sa_iterations}{Iteration number for SA (default = 10000)}

\item{ga_population_size}{Population size for GA (default = 400)}

\item{ga_iterations}{Iteration number for GA (default = 200)}

\item{ga_crossover_rate}{Applies crossover with the given probability in GA (default = 1, i.e. always perform crossover)}

\item{ga_mutation_rate}{For GA, applies mutation with given mutation rate (default = 0, i.e. mutation off)}

\item{gr_max_depth}{Sets max depth in greedy search, 0 for no limit (default = 1)}

\item{gr_search_depth}{Search depth in greedy search (default = 1)}

\item{gr_overlap_threshold}{Overlap threshold for results of greedy search (default = 0.5)}

\item{gr_subnetwork_num}{Number of subnetworks to be presented in the results (default = 1000)}
}
\value{
A list of genes in every identified active subnetwork that has a score greater than
the `score_quan_thr`th quantile and that has at least `sig_gene_thr` affected genes.
}
\description{
Performs active subnetwork search and filters identified active subnetworks
before returning final subnetworks.
}
\examples{
\donttest{
processed_df <- example_pathfindR_input[1:15, -2]
colnames(processed_df) <- c("GENE", "P_VALUE")
GR_snws <- get_active_subnetworks(processed_df, pin_name_path = "KEGG")
}
}
