% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_all_genes}
\alias{go_all_genes}
\title{Gene Ontology - All Gene Ontology Gene Sets}
\format{list containing 14586 vectors of gene symbols. Each vector corresponds
  to a gene set.}
\usage{
go_all_genes
}
\description{
A list containing the genes involved in each GO ontology term. Each
element is a vector of gene symbols located in the given gene set.
\emph{Generated on Sep 30, 2019.}
}
\keyword{datasets}
