% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_processing.R
\name{input_testing}
\alias{input_testing}
\title{Input Tesing}
\usage{
input_testing(input, p_val_threshold)
}
\arguments{
\item{input}{the input data that pathfindr uses. The input must be a data
frame with three columns: 1. Gene Symbol (HGNC Gene Symbol) 2. Change
value, e.g. log(fold change) 3. adjusted p value associated with test, e.g.
differential expression/methylation}

\item{p_val_threshold}{the adjusted-p value threshold to use when filtering
the input data frame. Must a numeric value between 0 and 1.}
}
\value{
Only checks if the input and the threshold follows the required
  specifications
}
\description{
Input Tesing
}
\examples{
input_testing()
}
