% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-visualization.R
\name{order_df_by_columnn}
\alias{order_df_by_columnn}
\title{Order input data frame by provided columnn}
\usage{
order_df_by_columnn(df, order_by)
}
\arguments{
\item{df}{the input data frame to be ordered}

\item{order_by}{A column name}
}
\value{
The ordered data frame or raises error
}
\description{
Order input data frame by provided columnn
}
