% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfindr.R
\docType{package}
\name{pathfindR}
\alias{pathfindR}
\title{pathfindR: A package for Enrichment Analysis Utilizing Active Subnetworks}
\description{
pathfindR is a tool for active-subnetwork-oriented gene set enrichment analysis.
The main aim of the package is to identify active subnetworks in a
protein-protein interaction network using a user-provided list of genes
and associated p values then performing enrichment analyses on the identified
subnetworks, discovering enriched terms (i.e. pathways, gene ontology, TF target
gene sets etc.) that possibly underlie the phenotype of interest.
}
\details{
For analysis on non-Homo sapiens organisms, pathfindR offers utility functions
for obtaining organism-specific PIN data and organism-specific gene sets data.

pathfindR also offers functionalities to cluster the enriched terms and
identify representative terms in each cluster, to score the enriched terms
per sample and to visualize analysis results.
}
\seealso{
See \code{\link{run_pathfindR}} for details on the pathfindR
active-subnetwork-oriented enrichment analysis
See \code{\link{cluster_enriched_terms}} for details on methods of enriched
terms clustering to define clusters of biologically-related terms
See \code{\link{score_terms}} for details on agglomerated score calculation
for enriched terms to investigate how a gene set is altered in a given sample
(or in cases vs. controls)
See \code{\link{term_gene_heatmap}} for details on visualization of the heatmap
of enriched terms by involved genes
See \code{\link{term_gene_graph}} for details on visualizing terms and
term-related genes as a graph to determine the degree of overlap between the
enriched terms by identifying shared and/or distinct significant genes
See \code{\link{UpSet_plot}} for details on creating an UpSet plot of the
enriched terms.
See \code{\link{get_pin_file}} for obtaining organism-specific PIN data and
\code{\link{get_gene_sets_list}} for obtaining organism-specific gene sets data
}
