% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RA_exp_mat}
\alias{RA_exp_mat}
\title{Example Input for pathfindR - Enriched Term Scoring}
\format{
A matrix with 572 rows and 33 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE15573}
}
\usage{
RA_exp_mat
}
\description{
A matrix containing the \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}}-transformed
and quantile-normalized expression values of the differentially-expressed
genes for 18 rheumatoid arthritis (RA) patients and 15 healthy subjects. The
matrix contains expression values of  572 significantly
differentially-expressed genes (see \code{\link{RA_input}}) with adj.P.Val <= 0.05.
\emph{Generated on Sep 28, 2019.}
}
\seealso{
\code{\link{RA_input}}  for the RA differentially-expressed genes data frame
\code{\link{RA_output}} for the RA example pathfindR enrichment output
\code{\link{score_terms}} for details on calculating agglomerated scores of enriched terms
}
\keyword{datasets}
