% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RA_output}
\alias{RA_output}
\title{Example Output for the pathfindR Enrichment Workflow - Rheumatoid Arthritis}
\format{A data frame with 104 rows and 8 columns:
\describe{
  \item{ID}{ID of the enriched term}
  \item{Term_Description}{Description of the enriched term}
  \item{Fold_Enrichment}{Fold enrichment value for the enriched term}
  \item{occurrence}{the number of iterations that the given term was found to enriched over all iterations}
  \item{lowest_p}{the lowest adjusted-p value of the given term over all iterations}
  \item{highest_p}{the highest adjusted-p value of the given term over all iterations}
  \item{Up_regulated}{the up-regulated genes in the input involved in the given term, comma-separated}
  \item{Down_regulated}{the down-regulated genes in the input involved in the given term, comma-separated}
}}
\usage{
RA_output
}
\description{
A dataset containing the results of pathfindR's active-subnetwork-oriented
enrichment workflow performed on the rheumatoid arthritis
differential-expression result dataset \code{RA_input}. Analysis via
\code{run_pathfindR} was performed using the default settings.
\emph{Generated on Oct 22, 2019.}
}
\seealso{
\code{\link{RA_input}} for example input of the enrichment workflow.
\code{\link{RA_clustered}} for example output of the clustering workflow.
\code{\link{run_pathfindR}} for details on the patfindR enrichment workflow.
}
\keyword{datasets}
