% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{return_pin_path}
\alias{return_pin_path}
\title{Return The Path to Given Protein Interaction Network (PIN)}
\usage{
return_pin_path(pin_name = "IntAct")
}
\arguments{
\item{pin_name}{Name of the chosen PIN. Must be one of c("Biogrid", "STRING",
"GeneMania", "BioPlex", "HitPredict", "IntAct"). Defaults to "IntAct".}
}
\value{
A character value that contains the path to chosen PIN.
}
\description{
Return The Path to Given Protein Interaction Network (PIN)
}
\examples{
pin_path <- return_pin_path("Biogrid")
}
\seealso{
See \code{\link{run_pathfindr}} for the wrapper function of the
  pathfindr workflow
}
