% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_pathfindr}
\alias{run_pathfindr}
\title{Wrapper Function for pathfindr Workflow}
\usage{
run_pathfindr(input, p_val_threshold = 0.05, enrichment_threshold = 1e-04,
  adj_method = "bonferroni", search_method = "GR",
  use_all_positives = FALSE, saTemp0 = 1, saTemp1 = 0.01,
  saIter = 10000, gaPop = 400, gaIter = 10000, gaThread = 5,
  gaMut = 0, grMaxDepth = 2, grSearchDepth = 1, grOverlap = 0.5,
  grSubNum = 1000, iterations = 10, n_processes = NULL,
  kegg_update = FALSE, pin_name_path = "GeneMania")
}
\arguments{
\item{input}{the input data that pathfindr uses. The input must be a data
frame with three columns: \enumerate{ \item Gene Symbol (HGNC Gene
Symbol) \item Change value, e.g. log(fold change) \item adjusted p value
associated with test, e.g. differential expression/methylation }}

\item{p_val_threshold}{the adjusted-p value threshold to use when filtering
the input data frame. Must a numeric value between 0 and 1.}

\item{enrichment_threshold}{threshold used when filtering individual pathway
enrichment results}

\item{adj_method}{correction method to be used for adjusting p-values of
pathway enrichment results (Default: 'bonferroni')}

\item{search_method}{algorithm to use when performing active subnetwork
search. Options are greedy search (GR), simulated annealing (SA) or genetic
algorithm (GA) for the search (Default:GR. Can be one of c("GR", "SA",
"GA"))}

\item{use_all_positives}{if TRUE in GA, adds an individual with all positive
nodes. In SA, initializes candidate solution with all positive nodes.
(Default = FALSE)}

\item{saTemp0}{initial temperature for SA (Default: 1.0)}

\item{saTemp1}{final temperature for SA (Default: 0.01)}

\item{saIter}{iteration number for SA (Default: 10000)}

\item{gaPop}{population size for GA (Default: 400)}

\item{gaIter}{iteration number for GA (Default: 10000)}

\item{gaThread}{number of threads to be used in GA (Default: 5)}

\item{gaMut}{the mutation rate for GA (Default: 0)}

\item{grMaxDepth}{sets max depth in greedy search, 0 for no limit (Default: 2)}

\item{grSearchDepth}{sets search depth in greedy search (Default: 1)}

\item{grOverlap}{sets overlap threshold for results of greedy search (Default:
0.5)}

\item{grSubNum}{sets number of subnetworks to be presented in the results
(Default: 1000)}

\item{iterations}{number of iterations for active subnetwork search and
enrichment analyses (Default = 10. Gets set to 1 for Genetic Algorithm)}

\item{n_processes}{optional argument for specifying the number of processes
used by foreach. If not specified, the function determines this
automatically (Default == NULL. Gets set to 1 for Genetic Algorithm)}

\item{kegg_update}{boolean value indicating whether or not to update KEGG
pathway genes (default = FALSE)}

\item{pin_name_path}{Name of the chosen PIN or path/to/PIN.sif. If PIN name,
must be one of c("Biogrid", "GeneMania", "IntAct", "KEGG"). If
path/to/PIN.sif, the file must comply with the PIN specifications. Defaults
to "GeneMania".}
}
\value{
Data frame of pathview enrichment results. Columns are: "ID",
 "Pathway", "occurence", "lowest_p", "highest_p". "ID" is the KEGG ID for a
 given pathway and "Pathway" is the name. "occurence" indicates the number of
 times the given pathway is encountered over all iterations. "lowest_p" and
 "highest_p" indicate the lowest and highest adjusted p values over all
 iterations. The function also creates an HTML report with the pathview
 enrichment results linked to the visualizations of the pathways in addition
 to the table of converted gene symbols. This report can be found in
 "results.html".
}
\description{
\code{run_pathfindr} is the wrapper function for the pathfindr workflow
}
\details{
This function takes in a data frame consisting of Gene Symbol, log-fold-change
and adjusted-p values. After input testing, any gene symbols that are not in
the PIN are converted to alias symbols if the alias is in the PIN. Next,
active subnetwork search is performed. Pathway enrichment analysis is
performed using the genes in each of the active subnetworks. Pathways with
adjusted-p values lower than \code{enrichment_threshold} are discarded. The
lowest adjusted-p value (over all subnetworks) for each pathway is kept. This
process of active subnetwork search and enrichment is repeated  for a selected
number of \code{iterations}, which is done in parallel. Over all iterations,
the lowest and the highest adjusted-p values, as well as number of occurences
are reported for each enriched pathway.
}
\section{Warning}{
 Depending on the protein interaction network of your choice,
 active subnetwork finding component of pathfindr may take a very long time
 to finish. Therefore, overnight runs are recommended.
}

\examples{
\dontrun{
run_pathfindr(RA_input)
}
}
\seealso{
\code{\link{input_testing}} for input testing,
 \code{\link{input_processing}} for input processing,
 \code{\link{current_KEGG}} for KEGG pathway genes retrieval,
 \code{\link{parseActiveSnwSearch}} for parsing a jActive modules output,
 \code{\link{enrichment}} for pathway enrichment analysis and
 \code{\link{pathmap}} for annotation of involved genes and visualization of
 pathways. See \code{\link[foreach]{foreach}} for details on parallel
 execution of looping constructs. See \code{\link{choose_clusters}} for
 clustering the resulting enriched pathways.
}
