% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_snw_search_and_enrichment_wrapper.R
\name{single_iter_wrapper}
\alias{single_iter_wrapper}
\title{Active Subnetwork Search + Enrichment Analysis Wrapper for a Single Iteration}
\usage{
single_iter_wrapper(
  i = NULL,
  pin_path,
  network,
  experiment_df,
  score_quan_thr,
  sig_gene_thr,
  search_method,
  verbose,
  start_with_all_positives,
  gene_init_prob,
  sa_initial_temp,
  sa_final_temp,
  sa_iterations,
  ga_population_size,
  ga_iterations,
  ga_crossover_rate,
  ga_mutation_rate,
  gr_max_depth,
  gr_search_depth,
  gr_overlap_threshold,
  gr_subnetwork_num,
  gset_list,
  adj_method,
  enrichment_threshold,
  list_active_snw_genes
)
}
\arguments{
\item{i}{current iteration index (default = \code{NULL})}

\item{pin_path}{path/to/PIN/file}

\item{network}{Prebuilt network object as returned by \code{build_network()},
built once by \code{active_snw_enrichment_wrapper} and passed to every
iteration to avoid redundant PIN file I/O.}

\item{score_quan_thr}{active subnetwork score quantile threshold. Must be
between 0 and 1 or set to -1 for not filtering. (Default = 0.8)}

\item{sig_gene_thr}{threshold for the minimum proportion of significant genes in
the subnetwork (Default = 0.02) If the number of genes to use as threshold is
calculated to be < 2 (e.g. 50 signif. genes x 0.01 = 0.5), the threshold number
is set to 2}

\item{search_method}{algorithm to use when performing active subnetwork
search. Options are greedy search (GR), simulated annealing (SA) or genetic
algorithm (GA) for the search (default = 'GR').}

\item{verbose}{boolean value indicating whether to print messages (default=FALSE)}

\item{start_with_all_positives}{if TRUE: in GA, adds an individual with all positive
nodes. In SA, initializes candidate solution with all positive nodes. (default = FALSE)}

\item{gene_init_prob}{For SA and GA, probability of adding a gene in initial solution (default = 0.1)}

\item{sa_initial_temp}{Initial temperature for SA (default = 1.0)}

\item{sa_final_temp}{Final temperature for SA (default = 0.01)}

\item{sa_iterations}{Iteration number for SA (default = 10000)}

\item{ga_population_size}{Population size for GA (default = 400)}

\item{ga_iterations}{Iteration number for GA (default = 200)}

\item{ga_crossover_rate}{Applies crossover with the given probability in GA (default = 1, i.e. always perform crossover)}

\item{ga_mutation_rate}{For GA, applies mutation with given mutation rate (default = 0, i.e. mutation off)}

\item{gr_max_depth}{Sets max depth in greedy search, 0 for no limit (default = 1)}

\item{gr_search_depth}{Search depth in greedy search (default = 1)}

\item{gr_overlap_threshold}{Overlap threshold for results of greedy search (default = 0.5)}

\item{gr_subnetwork_num}{Number of subnetworks to be presented in the results (default = 1000)}

\item{gset_list}{list for gene sets.}

\item{adj_method}{correction method to be used for adjusting p-values.
(default = 'bonferroni')}

\item{enrichment_threshold}{adjusted-p value threshold used when filtering
enrichment results (default = 0.05)}

\item{list_active_snw_genes}{boolean value indicating whether or not to report
the non-significant active subnetwork genes for the active subnetwork which was enriched for
the given term with the lowest p value (default = \code{FALSE})}
}
\value{
Data frame of enrichment results using active subnetwork search results
}
\description{
Active Subnetwork Search + Enrichment Analysis Wrapper for a Single Iteration
}
